requirements.txt
azure-functions==1.18.0
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
Flask==3.0.2
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pymongo==4.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
Werkzeug==3.0.1
